globals: size DAT_006b90f8 as a pointer, not a byte
FUN_004f3a00 does _DAT_006b90f8 = FUN_004f3980() (which returns void*) then
writes through *(_DAT_006b90f8 + 8). DAT_006b90f8 was declared uint8_t, so the
pointer was truncated to its low byte and the follow-up write faulted. Its
sibling pool (DAT_00792190, set the same way by FUN_004f3e40/FUN_004f3dc0) is
correctly a pointer; match that. Verified under wine: InitDirectDraw now runs
past FUN_004f3a00 to its final SetupZBuffer step.

+++ include/globals.h
@@ -538,7 +538,7 @@ extern uint8_t DAT_006aa66e;
 extern uint8_t DAT_006aa674;
 extern uint8_t DAT_006aa67c;
 extern uint8_t DAT_006aab04;
-extern uint8_t DAT_006b90f8;
+extern uint32_t DAT_006b90f8;
 #define _DAT_006b90f8 DAT_006b90f8
 extern uint32_t DAT_006b9100;
 extern uint8_t DAT_006ba1fc;
